See the individual commits for more information (#161)
* Updates to the response to the event

It's confusing to have two examples doing the same thing. The second example has too much abstraction to be useful.
I moved the first example into the right section as it is the payment finalization field.
Always a good idea to say if a function is pseudo code as I was unsure if that code was code sent in the initial adyen scripts for evaluation.
Parameters sent into `/payment` were completely off?!?

* Missed the undefined variable data

Changed data to the more likely `response` variable.